+++ app/src/main/java/com/knu/quickthink/components/CardComponent.kt
@@ -24,7 +24,7 @@ import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import com.google.accompanist.flowlayout.FlowRow
 import com.knu.quickthink.R
-import com.knu.quickthink.model.Card
+import com.knu.quickthink.model.card.Card
 
 
 @Composable
