+++ app/src/main/java/com/knu/quickthink/model/user/GoogleUserModel.kt
@@ -1,4 +1,4 @@
-package com.knu.quickthink.model
+package com.knu.quickthink.model.user
 
 import android.os.Parcelable
 import kotlinx.parcelize.Parcelize
