+++ app/src/main/java/com/knu/quickthink/model/user/TokenValidResponse.kt
@@ -1,4 +1,4 @@
-package com.knu.quickthink.model
+package com.knu.quickthink.model.user
 
 data class TokenValidResponse(
     val isValid : Boolean
