+++ app/src/main/java/com/knu/quickthink/repository/user/UserRepository.kt
@@ -1,11 +1,8 @@
-package com.knu.quickthink.repository
+package com.knu.quickthink.repository.user
 
 import com.google.android.gms.auth.api.signin.GoogleSignInAccount
-import com.knu.quickthink.UserToken
-import com.knu.quickthink.model.GoogleUserModel
-import com.knu.quickthink.model.IntroductionResponse
+import com.knu.quickthink.model.user.IntroductionResponse
 import com.knu.quickthink.model.NetworkResult
-import kotlinx.coroutines.flow.Flow
 
 interface UserRepository {
     suspend fun login(gsa :GoogleSignInAccount): NetworkResult<String>
