+++ app/src/main/java/com/knu/quickthink/screens/main/AccountViewModel.kt
@@ -5,7 +5,7 @@ import androidx.lifecycle.viewModelScope
 import com.knu.quickthink.model.onError
 import com.knu.quickthink.model.onException
 import com.knu.quickthink.model.onSuccess
-import com.knu.quickthink.repository.UserRepository
+import com.knu.quickthink.repository.user.UserRepository
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.StateFlow
