+++ app/src/main/java/com/knu/quickthink/screens/main/MainViewModel.kt
@@ -3,7 +3,7 @@ package com.knu.quickthink.screens.main
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import com.google.android.gms.auth.api.signin.GoogleSignInClient
-import com.knu.quickthink.repository.UserRepository
+import com.knu.quickthink.repository.user.UserRepository
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.StateFlow
